+++ Packages/com.unity.inputsystem/InputSystem/Plugins/XInput/XInputSupport.cs
@@ -25,9 +25,17 @@ namespace UnityEngine.InputSystem.XInput
                 matches: new InputDeviceMatcher().WithInterface("XInput"));
 #endif
 #if UNITY_STANDALONE_OSX || UNITY_EDITOR_OSX
+            // Legacy support when a user is using the 360Controller driver on macOS <= 10.15
             InputSystem.RegisterLayout<XboxGamepadMacOS>(
                 matches: new InputDeviceMatcher().WithInterface("HID")
                     .WithProduct("Xbox.*Wired Controller"));
+
+
+            // Matches macOS native support for Xbox Controllers
+            // macOS reports all Xbox controllers as "Controller" with manufacter Microsoft
+            InputSystem.RegisterLayout<XboxGamepadMacOSNative>(
+                matches: new InputDeviceMatcher().WithInterface("HID")
+                    .WithProduct("Controller").WithManufacturer("Microsoft"));
 
             // Matching older Xbox One controllers that have different View and Share buttons than the newer Xbox Series
             // controllers.